requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
python-crontab==2.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
